Move storage to seperate module
